+++ timevortex/settings/production.py
@@ -9,7 +9,7 @@ from timevortex.settings.base import *  # noqa
 #####
 # Logging configuration
 #
-LOG_BASE_FOLDER = "../common/log"
+LOG_BASE_FOLDER = '../common/log'
 LOGGING['handlers']['file']['filename'] = '%s/timevortex.log' % LOG_BASE_FOLDER  # noqa
 LOGGING['handlers']['file_weather']['filename'] = '%s/timevortex_weather.log' % LOG_BASE_FOLDER  # noqa
 LOGGING['handlers']['file_energy']['filename'] = '%s/timevortex_energy.log' % LOG_BASE_FOLDER  # noqa
@@ -21,7 +21,7 @@ LOGGING['loggers']['energy']['level'] = 'DEBUG'  # noqa
 # Timevortex configuration
 #
 
-SETTINGS_FILE_STORAGE_FOLDER = "../common/data"
+SETTINGS_FILE_STORAGE_FOLDER = '../common/data'
 
 #####
 # Email configuration
@@ -39,3 +39,9 @@ EMAIL_USE_TLS = True
 #
 
 BACKUP_TARGET_FOLDER = None
+
+#####
+# DATABASE CONFIGURATION
+#
+
+DATABASES['default']['NAME'] = '../common/db.sqlite3'  # noqa